scripts/requirements.txt
fasttext-wheel==0.9.2
numpy<2.0.0
